domain/comment/abuse table population
