--- app/src/main/java/za/ac/cput/inforshare/repository/db/model/UserStatus.java
@@ -1,8 +1,0 @@
-package za.ac.cput.inforshare.repository.db.model;
-
-/**
- * Created by hashcode on 2018/02/11.
- */
-
-public class UserStatus {
-}
