+++ app/src/main/java/za/ac/cput/inforshare/repository/db/model/users/UserLanguage.java
@@ -1,4 +1,4 @@
-package za.ac.cput.inforshare.repository.db.model;
+package za.ac.cput.inforshare.repository.db.model.users;
 
 import org.greenrobot.greendao.annotation.Entity;
 import org.greenrobot.greendao.annotation.Generated;
@@ -6,7 +6,7 @@ import org.greenrobot.greendao.annotation.Id;
 import org.greenrobot.greendao.annotation.Property;
 
 /**
- * Created by christian on 2018/02/13.
+ * Created by hashcode on 2018/02/13.
  */
 
 @Entity(nameInDb = "UserLanguage")
